fixing importand database problem, table delete with no table to delete

+++ db/migrate/20230802165403_remove_system_from_parts.rb
@@ -1,5 +1,6 @@
 class RemoveSystemFromParts < ActiveRecord::Migration[7.0]
   def change
-    remove_column :parts, :system, :string
+    remove_column :parts, :system, :string if column_exists?(:parts, :system)
+   
   end
 end
