+++ db/migrate/20230802165420_remove_category_from_parts.rb
@@ -1,5 +1,5 @@
 class RemoveCategoryFromParts < ActiveRecord::Migration[7.0]
   def change
-    remove_column :parts, :category, :string
+    remove_column :parts, :category, :string if column_exists?(:parts, :category)
   end
 end
