+++ db/migrate/20230816083106_drop_order_storage_lists.rb
@@ -1,7 +1,10 @@
 class DropOrderStorageLists < ActiveRecord::Migration[7.0]
   def up
-    drop_table :order_storage_lists
+    if table_exists? :order_storage_lists
+      drop_table :order_storage_lists
+    end
   end
+  
 
   def down
     raise ActiveRecord::IrreversibleMigration
